Added a config option to not delete out files and populated the results dict.

+++ helper_tools/files.py
@@ -24,6 +24,32 @@ def read_file(file):
         return unicode_error_msg
 
 
+def initialize_results():
+    key_list = []
+    stud_dict = {}
+    navi = helper_tools.navigation.Dirs()
+
+    # get key files
+    for file in os.listdir(navi.key_dir):
+        if file.endswith('.py'):
+            try:
+                source = open(file, 'r').read()
+            except UnicodeDecodeError:
+                source = unicode_error_msg
+            key_list.append({'source': source, 'file name': file})
+
+    # get student files
+    for stud_name in navi.students:
+        stud_dict[stud_name] = []
+        for file in os.listdir(join(navi.result_dir, stud_name)):
+            if file.endswith('.py'):
+                try:
+                    source = open(file, 'r').read()
+                except UnicodeDecodeError:
+                    source = unicode_error_msg
+                stud_dict[stud_name].append({'source': source, 'file name': file})
+
+
 def run_files(dicts):
     base_threads = threading.active_count()
     if config.max_concurrent_programs == 0:
@@ -37,7 +63,7 @@ def run_files(dicts):
             dict_obj = dicts[num_run]
             threads.append(threading.Thread(target=run_a_file,
                                             args=(join(dict_obj['file path'], dict_obj['file name']),
-                                                  config.file_out_name)))
+                                                  dict_obj['file name'] + config.file_out_name)))
             threads[-1].start()
 
     for thread in threads:
@@ -45,7 +71,8 @@ def run_files(dicts):
 
     for dict_obj in dicts:
         dict_obj['out'] = read_file(join(dict_obj['file path'], config.file_out_name))
-        os.remove(join(dict_obj['file path'], config.file_out_name))
+        if config.del_ouput_files:
+            os.remove(join(dict_obj['file path'], config.file_out_name))
 
 
 def run_key():
